requirements.txt
boto3==1.24.12
lmfit==1.0.3
matplotlib==3.5.2
numpy==1.20.1
pandas==1.2.3
poolmanager==0.2.1
qiskit==0.36.2
qiskit_ionq==0.3.3
qiskit_terra==0.20.2
qutip==4.7.0
response==0.5.0
scipy==1.6.1
seaborn==0.11.2
sympy==1.7.1
tweedledum==1.1.1
amazon-braket-sdk==1.25.1
pylatexenc==2.10